+++ src/JsonApi/Exception/JsonApiExceptionInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 namespace WoohooLabs\Yin\JsonApi\Exception;
 
 use Throwable;
-use WoohooLabs\Yin\JsonApi\Schema\Document\AbstractErrorDocument;
+use WoohooLabs\Yin\JsonApi\Schema\Document\ErrorDocumentInterface;
 
 interface JsonApiExceptionInterface extends Throwable
 {
-    public function getErrorDocument(): AbstractErrorDocument;
+    public function getErrorDocument(): ErrorDocumentInterface;
 }