+++ src/JsonApi/Exception/ResourceIdentifierIdInvalid.php
@@ -5,7 +5,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 
 use WoohooLabs\Yin\JsonApi\Schema\Error\Error;
 
-class ResourceIdentifierIdInvalid extends JsonApiException
+class ResourceIdentifierIdInvalid extends AbstractJsonApiException
 {
     /**
      * @var string