+++ src/JsonApi/Exception/SortParamUnrecognized.php
@@ -6,7 +6,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 use WoohooLabs\Yin\JsonApi\Schema\Error\Error;
 use WoohooLabs\Yin\JsonApi\Schema\Error\ErrorSource;
 
-class SortParamUnrecognized extends JsonApiException
+class SortParamUnrecognized extends AbstractJsonApiException
 {
     /**
      * @var string